requirements.txt
asgiref==3.9.1
crispy-bootstrap5==0.7
Django==5.2.5
django-crispy-forms==2.1
pillow==11.0.0
python-slugify==8.0.1
setuptools==80.9.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
wheel==0.45.1
